+++ pytorch_lightning/loops/dataloader/dataloader_loop.py
@@ -56,9 +56,9 @@ class DataLoaderLoop(Loop):
     def reset(self) -> None:
         """Resets the internal state."""
         if not self.restarting:
-            self.dataloader_progress.current.reset()
+            self.dataloader_progress.reset_on_run()
         else:
-            self.dataloader_progress.current.reset_on_restart()
+            self.dataloader_progress.reset_on_restart()
 
     def on_advance_start(self, *args: Any, **kwargs: Any) -> None:
         self.dataloader_progress.increment_ready()