+++ pytorch_lightning/utilities/__init__.py
@@ -41,7 +41,6 @@ from pytorch_lightning.utilities.imports import (  # noqa: F401
     _IS_INTERACTIVE,
     _JSONARGPARSE_AVAILABLE,
     _module_available,
-    _NATIVE_AMP_AVAILABLE,
     _OMEGACONF_AVAILABLE,
     _POPTORCH_AVAILABLE,
     _RICH_AVAILABLE,