+++ tests/loggers/test_tensorboard.py
@@ -120,6 +120,7 @@ def test_tensorboard_no_name(tmpdir, name):
     assert os.listdir(tmpdir / "version_0")
 
 
+@mock.patch.dict(os.environ, {})
 def test_tensorboard_log_sub_dir(tmpdir):
     class TestLogger(TensorBoardLogger):
         # for reproducibility